Separate query and server assertion
