+++ src/Query/Mixin/Returning.php
@@ -2,7 +2,7 @@
 
 namespace Misantron\QueryBuilder\Query\Mixin;
 
-use Misantron\QueryBuilder\Assert\Assert;
+use Misantron\QueryBuilder\Assert\QueryAssert;
 
 /**
  * Trait Returning.
@@ -23,7 +23,7 @@ trait Returning
      */
     public function returning($items)
     {
-        Assert::columnsNotEmpty($items);
+        QueryAssert::columnsNotEmpty($items);
 
         $this->returning = $this->parseList($items);
